+++ packages/components/src/grid-table/demo/虚拟列表.preview.vue
@@ -8,8 +8,7 @@
 </template>
 
 <script lang="tsx" setup>
-  import { MixteGridTable } from '@mixte/components/grid-table';
-  import { defineTableColumns } from '@mixte/components/grid-table/utils';
+  import { defineTableColumns, MixteGridTable } from '@mixte/components/grid-table';
   import { ElButton } from 'element-plus';
   import { randomNatural } from 'mixte';
   import '@mixte/components/dist/grid-table/css/index.scss';