+++ packages/components/src/grid-table/src/components/empty.vue
@@ -23,7 +23,6 @@
   function RenderSvg() {
     return (
       <svg width="64" height="41" viewBox="0 0 64 41" xmlns="http://www.w3.org/2000/svg">
-        <title>暂无数据</title>
         <g transform="translate(0 1)" fill="none" fill-rule="evenodd">
           <ellipse fill="#f5f5f5" cx="32" cy="33" rx="32" ry="7" />
           <g fill-rule="nonzero" stroke="#d9d9d9">