+++ packages/components/src/grid-table/src/composables/useShared.ts
@@ -71,9 +71,9 @@ export const [
         // 未设置
         if (width == null) gridTemplateColumns += 'auto ';
         // 数字宽度
-        else if (isNumeric(width)) gridTemplateColumns += `${width}px `;
+        else if (isNumeric(width) && Number(width) >= 0) gridTemplateColumns += `${width}px `;
         // 合法的宽度
-        else if (CSS.supports('grid-template-columns', width)) gridTemplateColumns += `${width} `;
+        else if (CSS.supports('grid-template-columns', `${width} auto`)) gridTemplateColumns += `${width} `;
         // 其他情况
         else gridTemplateColumns += 'auto ';
       }