+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - 💄 [@mixte/components] 组件 `GridTable` 支持展开树形数据功能 *
   - 💄 [@mixte/components] 优化 `GridTable` 组件虚拟列表的动态高度计算相关逻辑 *
   - 💄 [@mixte/use] 优化 `useRequest` 的 `UseRequestOptions` 类型定义 *
+  - 💄 [@mixte/components] 组件 `GridTable` 默认导出新增 `defineTableColumns` 方法导出 *
 
 ## [v3.5.0-beta.3]
   - 📅 2025-06-19